Added a field in the SHPImportWizard to allow users to enter the projection SRS code of the shape file. This field is put into the SQLConfig.xml after the Shape file is parsed.
You will need to redeploy the Java project.

Tested with Adminconsole.xml by importing two sets of shape files with
projection SRS codes and verifying the codes are correct inside the
Geometry Collections editor and in an instance of weave. Also tested
saving changes to an geometry collection still works.

Change-Id: I3e87bdb35c88ab22b37bb78f381eb9c173273e31