Added interface ISpatialIndexImplementation.
Added static class SpatialIndexFactory whose purpose is to provide an
ISpatialIndexImplementation object depending on the type of plotter.

Added GeometrySpatialIndex and RefinedSpatialIndex which are
ISpatialIndexImplementation objects.

Modified SpatialIndex to be a bridge class for the
ISpatialIndexImplementation classes. Essentially, SpatialIndex
forwards the requests in most cases, but handles some requests.

Changed the probing buffer back to 10 screen pixels.

Tested with Obesity.xml to verify the behavior is the same as last
check in.

Change-Id: Id1c683b4881394222eec37b7695b14b3148edf27