Made ComputationalGeometryUtils use the more efficient lineIntersectsLine function.
Added an option in User Interface Settings to enable the geometry
probing which is much more accurate, but may hurt performance.

SpatialIndex will revert to the old behavior if the geometry probing
is not enabled or if the plotter isn't an IPlotterWithGeometries.

Tested with obesity.xml by probing the states and borders. Also
verified probing a bar chart, line chart, and scatter plot with
different keys works. Disabled the geometry probing and observed the
old behavior in the MapTool.

Change-Id: I6aa93a064dcdc24f1ce8fb92d24a916dbae95b25