Amending last commit.
Modified SpatialIndex to use exact geometry probing with a buffer
which works as expected. It either gets the exact key that overlaps or
then nearest key which intersects the buffer for all plotters.

Fixed zooming in the vis tools by using a clip rectangle to clip the
drawing of the shape to the screenBounds. This keeps the line chart
and compound bar chart (for example) from overlapping the axes.
Also trimmed the probing for these to the dataBounds to prevent
offscreen probing.

Tested with kyle.xml and obesity.xml by verifying selection and
probing works as expected.

Change-Id: Ibbd8b71ab211d2c24d539ab2c5013fbcd06bc0d0