Fixed small issues with SpatialIndex. The index now considers lines and points very close to the center of the bounds as overlapping the query center.
Added a WeaveProperty for sessionStateEditor. This is changed every
time the user clicks the Toggle View button or changes it in the
editor.

Removed a lot of unnecessary code in the SessionStateEditor. The
DataGrid now looks directly at the ILinkableVariables and displays
them. It does not display any arrays or references. 

Tested everything with Obesity.xml by probing everything, selecting
everything, and messing around with the session state editor (editing,
reloading, saving, closing, toggling view).

Change-Id: I1a9e6afca9d364f63bedf22b6ab4781e9b0d0925