Disabled super user options for non super users.  Disabled overwrite for non super users as well.
LoginPopup no longer has a combobox of connections.

ForceMigrateWizard creates a connection which is always a superuser
now.

Added connectionName and password parameter to some ASync java calls
to verify connection and config.

Fixed a bug in getAttributeColumnInfo where a reference to an invalid
GeometryCollectionInfo would throw an exception. Instead, the server
ignores this is if it fails.

Change-Id: I1bd69047d4195507aa2e68ee7b361ec2cbae0b92