Renamed superUser/privileges to is_manager. Removed unnecessary filtering, fixed some issues with the get functions in the Java project at runtime.
Errors are thrown if a regular user attempts to overwrite a data table
or geometry collection outside of the connection.

Fixed bug with AdminConsole receiving an error when a user tried to
view the DataTable with an invalid GeometryCollection.

Change database connections is always enabled if the config is
considered migrated. There are now checks to never allow the user to
remove the last connection or the last manager account.

Tested with AdminConsole.xml and Kyle.xml. The former was used to
check every function of the AdminConsole still worked. The latter was
used to check the client side still functioned.

Change-Id: I32f3aaf7c8c2df87ac77badc9f167d9f8738feac