Added a DashedLine class which is used for drawing dashed lines with arbitrary dashes. This is used primarily for the rectangle drawn in during selection/zooming in a tool.
Added a way to modify the parameters in the User Interface Preferences
with the default option on. It also shows a preview of the setting.

Tested with obesity.xml on each  tool and verified session state and
the preferences panel all worked. Also verified that the stored is
never invalid (not all values are 0).

Change-Id: Iac6f7b704f72bfc7177dca18812aacc80e16fe9e