Fixed project source files setting where it included some graph files.
IWMSService interface now has an extra parameter to prefer
lowerQuality. The classes which implement or extend this have been
appropriately updated. This fixes the lower quality issue with the
WMSLayer Settings. Previously, this was handled by resizing data and
screen bounds, but that would request many tiles offscreen. The
parameter neatly handles it, and forces the OnEarthProvider and
ModestMapsWMS wrapper classes to handle it.

Tested with obesity.xml by adding a WMS layer, changing service, and
toggling the parameter in the layer's settings.

Change-Id: I7f20c774c4832818c6961cbd0cafe0c9daad8d3f