Commented out some println calls which were always printing in functions.
Added proper handling of types when parsing CSV data in
AdminService.importCSV(). Previously, the last row determined the type
which meant a column with a string in the first row could be an int or
double type. The new behavior keeps the column of type String or
Double if it was previously determined to be such.

Also added support for treating "integers" like 04 and 01 as strings. 

Tested by importing tables in the AdminConsole which had strings,
doubles, and ints in various positions in columns and verifying
correct handling.

Change-Id: I59043037a5f58392d82f2bc81a0fbfc5e1a62b0f