Cleaned up contextMenu code for DataTableTool and fixed bug where it would sometimes export the wrong CSV and crash.
Also fixed bug where exportCSV would return without doing any work if
it was called from the settings panel for the tool.

Tested by creating an xml with 8 DataTableTools, set each to have
different columns and tested export with right click for each and the
export data button in the settings panel. 

Change-Id: I772b064b23ae194cf8d16f790b22922aac9d67f7