Fixed bug where the spatialindex.createIndex() function would be called when probing/selecting. The reason for this is the spatial index had a record count of 0 (when the bounds array for the keys was empty), but the index was created previously.
When the index is created, a flag is set telling the plot layer that
it doesn't need to validate the index. This fixes the bug.

Tested with obesity-cube.xml by probing and selecting with a
breakpoint in createIndex and observing that it wasn't being called
except for validating the index.

Change-Id: I200e732045ef190366b11af02db02a1f5b4aacac